chg: test_gen_ai_utilities.py
