+++ docker/kafka/README.md
@@ -1,7 +1,6 @@
 # Kafka
 
-This is used as a message queue service to connect the checkout service with
-the accounting and fraud detection services.
+This is used as a message queue service to connect the checkout service with the accounting and fraud detection
+services.
 
-Kafka is run in KRaft mode. Environment variables are substituted at
-deploy-time.
+Kafka is run in KRaft mode. Environment variables are substituted at deploy-time.
